% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Andahuaylas.R
\docType{data}
\name{Andahuaylas}
\alias{Andahuaylas}
\title{A dataset of historical precipitation and information about Andahuaylas Station}
\format{
An object of class \code{data.frame} with 13354 rows and 17 columns.
}
\source{
data downloaded from ANA (Peru), these data being distributed without any warranty.
}
\usage{
Andahuaylas
}
\description{
A dataset of historical precipitation and information about Andahuaylas Station (ANA, Perú).
}
\keyword{datasets}
