% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa.info.R
\name{oa.info}
\alias{oa.info}
\title{function for reading, sorting and convert a data file (station information) downloaded from ANA in dataframe.}
\usage{
oa.info(x)
}
\arguments{
\item{x}{a file (xlsx) downloaded from ANA (Peru)}
}
\description{
function for reading, sorting and convert a data file (station information) downloaded from ANA in dataframe.
}
\author{
Geomar Perales Apaico
}
