% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oaxlsx.R
\name{oaxlsx}
\alias{oaxlsx}
\title{function for reading, sorting and convert a data file downloaded from ANA in dataframe.}
\usage{
oaxlsx(x)
}
\arguments{
\item{x}{a file (xlsx) downloaded from ANA (Peru)}
}
\description{
function for reading, sorting and convert a data file downloaded from ANA in dataframe.
}
\author{
Geomar Perales Apaico
}
